# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for PPU Smart Assistant
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface OtpRequest {
  email: string;
}

export interface OtpResponse {
  token: string;
  userId: string;
  email: string;
  studentName?: string | null;
  major?: string | null;
  isAuthenticated: boolean;
}

export interface OtpVerify {
  email: string;
  code: string;
}

export interface AuthSession {
  userId: string;
  email: string;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  major?: string | null;
  isAuthenticated: boolean;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: string;
  role: ChatMessageRole;
  content: string;
  imageUrl?: string | null;
  createdAt: string;
}

export interface ChatInput {
  /** @minLength 1 */
  message: string;
  imageBase64?: string | null;
  mimeType?: string | null;
}

export interface Student {
  id: string;
  name: string;
  email: string;
  major: string;
  enrolledSubjects?: string[];
}

export interface ScheduleEntry {
  id: string;
  subjectName: string;
  subjectCode: string;
  dayOfWeek: string;
  startTime: string;
  endTime: string;
  room?: string;
  doctorName?: string;
}

export interface Subject {
  id: string;
  name: string;
  code: string;
  major: string;
  level: number;
  creditHours: number;
  /** @nullable */
  description?: string | null;
}

export interface Doctor {
  id: string;
  name: string;
  /** @nullable */
  email?: string | null;
  subjects: string[];
  /** @nullable */
  department?: string | null;
  /** @nullable */
  officeHours?: string | null;
}

export interface ActivityItem {
  id: string;
  description: string;
  timestamp: string;
}

export interface DashboardStats {
  totalSubjects: number;
  totalDoctors: number;
  totalStudents: number;
  totalChatMessages: number;
  recentActivity: ActivityItem[];
}

export type GetChatHistoryParams = {
limit?: number;
};

export type ListSubjectsParams = {
major?: string;
level?: number;
};

export type ListDoctorsParams = {
subject?: string;
};