# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for PPU Smart Assistant
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: OTP-based authentication
  - name: chat
    description: AI chat operations
  - name: students
    description: Student profile and data
  - name: subjects
    description: University subjects
  - name: doctors
    description: University doctors/professors
  - name: dashboard
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request OTP via email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpResponse"
        "400":
          description: Invalid email or not registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and start session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerify"
      responses:
        "200":
          description: OTP verified, session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout and invalidate session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /chat/messages:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat message history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message to AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  /chat/messages/clear:
    delete:
      operationId: clearChatHistory
      tags: [chat]
      summary: Clear all chat history
      responses:
        "200":
          description: History cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /students/me:
    get:
      operationId: getStudentProfile
      tags: [students]
      summary: Get current student profile
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /students/schedule:
    get:
      operationId: getStudentSchedule
      tags: [students]
      summary: Get student weekly schedule
      responses:
        "200":
          description: Weekly schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleEntry"

  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all subjects
      parameters:
        - name: major
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"

  /subjects/{id}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get subject details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subject details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "404":
          description: Subject not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List all doctors/professors
      parameters:
        - name: subject
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"

  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get doctor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Doctor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "404":
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    OtpRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email

    OtpResponse:
      type: object
      properties:
        token:
          type: string
        userId:
          type: string
        email:
          type: string
        studentName:
          type: string
          nullable: true
        major:
          type: string
          nullable: true
        isAuthenticated:
          type: boolean
      required:
        - token
        - userId
        - email
        - isAuthenticated

    OtpVerify:
      type: object
      properties:
        email:
          type: string
          format: email
        code:
          type: string
      required:
        - email
        - code

    AuthSession:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        studentName:
          type: ["string", "null"]
        major:
          type: ["string", "null"]
        isAuthenticated:
          type: boolean
      required:
        - userId
        - email
        - isAuthenticated

    ChatMessage:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - role
        - content
        - createdAt

    ChatInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
        imageBase64:
          type: string
          nullable: true
        mimeType:
          type: string
          nullable: true
      required:
        - message

    Student:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        major:
          type: string
        enrolledSubjects:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - email
        - major

    ScheduleEntry:
      type: object
      properties:
        id:
          type: string
        subjectName:
          type: string
        subjectCode:
          type: string
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: string
        doctorName:
          type: string
      required:
        - id
        - subjectName
        - subjectCode
        - dayOfWeek
        - startTime
        - endTime

    Subject:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        major:
          type: string
        level:
          type: integer
        creditHours:
          type: integer
        description:
          type: ["string", "null"]
      required:
        - id
        - name
        - code
        - major
        - level
        - creditHours

    Doctor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        subjects:
          type: array
          items:
            type: string
        department:
          type: ["string", "null"]
        officeHours:
          type: ["string", "null"]
      required:
        - id
        - name
        - subjects

    DashboardStats:
      type: object
      properties:
        totalSubjects:
          type: integer
        totalDoctors:
          type: integer
        totalStudents:
          type: integer
        totalChatMessages:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required:
        - totalSubjects
        - totalDoctors
        - totalStudents
        - totalChatMessages
        - recentActivity

    ActivityItem:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - description
        - timestamp