# lib/api-zod/src/generated/types/otpResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for PPU Smart Assistant
 * OpenAPI spec version: 0.1.0
 */

export interface OtpResponse {
  token: string;
  userId: string;
  email: string;
  studentName?: string | null;
  major?: string | null;
  isAuthenticated: boolean;
}